+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-2495] Cached structures metadata in Structure Cache
+
 ## [3.20.2] 2020-04-14
 
 ### Added